use static addresses for host & backened interfaces
To remove calico dependencies we are going to use static addresses
for host interface.
We are aslo adding one other interace called 'backendInterface' which will
be used to hold the Pod GW address as a default route from Pod.

Both of these addresses are taken from variables in pkg/types/types.go.

Signed-off-by: Abdul Halim <abdul.halim@intel.com>